Fix Docker image name handling
- Make slashes consistent across all Makefiles.
- Rename `REPO` and `DOCKER_REPO` to `REGISTRY` and `DOCKER_REGISTRY`.
- Fix typos in comments/docs.

The current state has inconsistent handling of Docker image names.
Some of the images have slashes after the *registry* component in the
image name while others don't. Additionally, the variables which are
used to set the registry components are called "REPO" and
"DOCKER_REPO", which is confusing since the *repository* component of
a Docker image is actually the part after the first slash and before
the colon. The part which comes before the slash is the registry.