Fixes and improvements
* Refactor the checkClusterHealth to run before initiating the
reconcile.

This new step should considerably improve the workflow by checking the
cluster health before doing anything at all. If the cluster is not in an
ok state, the operator should create/update any additional resources and
should just wait until the cluster is "ok" again.

The additional benefit is that we don't need to check the health of the
cluster again during the specific StatefulSet reconcile call, which
significantly simplified the code.

* Refactor reconciliationContext to include all current stateful sets,
all pods and sidecar clients. Additionally it now includes the type of
the current sclaing operation (up, down or none)

This allows building these resources once for any given reconciliation
cycle. They also become available immediately to lots of other functions
that already receive this context and make the code more readable.

* Small fixes and improvements.
* More comments explaining important areas.

Signed-off-by: Alex Lourie <djay.il@gmail.com>