Re-add deleted manifests
We should keep the original manifests in addition to the "bundle"
YAML files we generate. The bundle is generated from the original
manifests. In addition, some manifests are automatically created by
the operator SDK (e.g. `operator.yaml`) so deleting them will create
problems when upgrading the operator to newer SDK versions.