Added more checks in CI
Additional "repo" checks in CI will validate that OpenAPI and k8s
objects are in sync with the code changes; as well as that `make bundle`
was rerun and is also in sync with the current code state.

Also, switch to a separate validation script vs having the code in
circle CI config.

Signed-off-by: Alex Lourie <djay.il@gmail.com>